gestione utenze completato

+++ src/main/java/it/prova/myebay/service/AnnuncioServiceImpl.java
@@ -9,10 +9,10 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import it.prova.myebay.exception.AnnuncioChiusoException;
+import it.prova.myebay.exception.UtenteNotFoundException;
 import it.prova.myebay.model.Annuncio;
 import it.prova.myebay.model.Utente;
 import it.prova.myebay.repository.annuncio.AnnuncioRepository;
-import it.prova.myebay.repository.utente.UtenteRepository;
 
 
 @Service
@@ -46,6 +46,18 @@ public class AnnuncioServiceImpl implements AnnuncioService{
 	@Override
 	@Transactional
 	public void aggiorna(Annuncio annuncioInstance) {
+		String username = SecurityContextHolder.getContext().getAuthentication().getName();
+		Utente utenteFromDb = utenteService.findByUsername(username);
+		
+		if (utenteFromDb == null)
+		throw new UtenteNotFoundException();
+		
+		if (!annuncioInstance.isAperto()) {
+			throw new AnnuncioChiusoException();
+		} 
+		annuncioInstance.setAperto(true);
+		annuncioInstance.setDataCreazione(LocalDate.now());
+		annuncioInstance.setUtente(utenteFromDb);
 		annuncioRepository.save(annuncioInstance);
 	}
 
@@ -56,7 +68,7 @@ public class AnnuncioServiceImpl implements AnnuncioService{
 		Utente utenteFromDb = utenteService.findByUsername(username);
 		
 		if (utenteFromDb == null)
-		throw new RuntimeException("Elemento non trovato.");
+		throw new UtenteNotFoundException();
 		
 		annuncioInstance.setUtente(utenteFromDb);
 		annuncioInstance.setAperto(true);
@@ -67,9 +79,16 @@ public class AnnuncioServiceImpl implements AnnuncioService{
 	@Override
 	@Transactional
 	public void rimuovi(Long idAnnuncio) {
+		String username = SecurityContextHolder.getContext().getAuthentication().getName();
+		Utente utenteFromDb = utenteService.findByUsername(username);
 		Annuncio annuncioFromDB= this.caricaSingoloElemento(idAnnuncio);
+		if (utenteFromDb == null)
+			throw new UtenteNotFoundException();
 		if (!annuncioFromDB.isAperto()) {
 			throw new AnnuncioChiusoException();
+		}
+		if (annuncioFromDB.getUtente() != utenteFromDb) {
+			throw new RuntimeException();
 		}
 		annuncioRepository.deleteById(idAnnuncio);
 	}
@@ -78,6 +97,11 @@ public class AnnuncioServiceImpl implements AnnuncioService{
 	@Transactional(readOnly = true)
 	public List<Annuncio> findByExampleRicerca(Annuncio example) {
 		return annuncioRepository.findByExampleRicerca(example);
+	}
+
+	@Override
+	public Annuncio caricaElementoConUtente(Long id) {
+		return annuncioRepository.findByIdConUtente(id).orElse(null);
 	}
 
 }
